Fix styled-components theme types for lint-staged
